Don't send out a chat marker for each MAM result
Fix this by adding a new message property `is_last_mam_result`
and only attempt to send out a chat marker if the message is archived
and that property is `true`.

Also adds a new property `archive_id` which is a duplicate of `stanza_id ${jid}`.
Makes it easier to get the archive id without having to look up the JID
of the archiving entity.